MY BSQ
Frankly spaeking I did everthing to make this code during the weeks :)). First of all I learned about what the task is itself then i tried to make algorith for it 
More I used my math knowledge to find biggest square and it is index to make int array to calculate minimum of three parts of square and so one 
So I open file y and read its characters one by one and stored them to one string called content then I check all of index of it 
with it I made integer matrix according to content of file so if i see '.' I stored 1 value to integer matrix elsif 'o' I stored 0 value to it
then with it I used dynamic algorithm method to calculate and find and define biggest square of map. When I check one index of map I take 
three sides of it . And I find minimum of them then I continue calculation . If there is no null namely 0 value , size of square that I am trying to find
must continue incresing . After I find size and index of biggets square of map I had to define another 3 sides of it after that I have to 
print old file but with showing best biggest square of map with 'x' .........